Make DialogHandler interface, extract Staged Security dialogs to a delegate class
