+++ app/src/main/java/com/tari/android/wallet/ui/fragment/home/overview/HomeOverviewViewModel.kt
@@ -3,8 +3,6 @@ package com.tari.android.wallet.ui.fragment.home.overview
 
 import android.app.Activity
 import android.os.Build
-import android.text.SpannableString
-import android.text.Spanned
 import androidx.lifecycle.MediatorLiveData
 import com.tari.android.wallet.R
 import com.tari.android.wallet.R.string.error_no_connection_description
@@ -14,16 +12,13 @@ import com.tari.android.wallet.R.string.error_node_unreachable_title
 import com.tari.android.wallet.application.deeplinks.DeepLink
 import com.tari.android.wallet.application.deeplinks.DeeplinkManager
 import com.tari.android.wallet.application.securityStage.StagedWalletSecurityManager
-import com.tari.android.wallet.application.securityStage.StagedWalletSecurityManager.StagedSecurityEffect
 import com.tari.android.wallet.application.walletManager.WalletManager.WalletEvent
 import com.tari.android.wallet.data.BalanceStateHandler
 import com.tari.android.wallet.data.TransactionRepository
 import com.tari.android.wallet.data.sharedPrefs.CorePrefRepository
-import com.tari.android.wallet.data.sharedPrefs.securityStages.WalletSecurityStage
 import com.tari.android.wallet.data.sharedPrefs.sentry.SentryPrefRepository
 import com.tari.android.wallet.extension.collectFlow
 import com.tari.android.wallet.extension.launchOnIo
-import com.tari.android.wallet.extension.takeIfIs
 import com.tari.android.wallet.model.BalanceInfo
 import com.tari.android.wallet.model.Tx
 import com.tari.android.wallet.navigation.Navigation
@@ -35,12 +30,8 @@ import com.tari.android.wallet.ui.dialog.modular.modules.body.BodyModule
 import com.tari.android.wallet.ui.dialog.modular.modules.button.ButtonModule
 import com.tari.android.wallet.ui.dialog.modular.modules.button.ButtonStyle
 import com.tari.android.wallet.ui.dialog.modular.modules.head.HeadModule
-import com.tari.android.wallet.ui.dialog.modular.modules.securityStages.SecurityStageHeadModule
-import com.tari.android.wallet.ui.dialog.modular.modules.space.SpaceModule
 import com.tari.android.wallet.ui.fragment.contactBook.data.ContactsRepository
 import com.tari.android.wallet.ui.fragment.send.finalize.TxFailureReason
-import com.tari.android.wallet.ui.fragment.settings.backup.backupOnboarding.item.BackupOnboardingArgs
-import com.tari.android.wallet.ui.fragment.settings.backup.backupOnboarding.module.BackupOnboardingFlowItemModule
 import com.tari.android.wallet.ui.fragment.tx.adapter.TransactionItem
 import com.tari.android.wallet.util.extractEmojis
 import com.tari.android.wallet.util.shortString
@@ -49,7 +40,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.withContext
-import yat.android.ui.extension.HtmlHelper
 import javax.inject.Inject
 
 
@@ -82,6 +72,13 @@ class HomeOverviewViewModel : CommonViewModel() {
         component.inject(this)
     }
 
+    private val stagedSecurityDelegate = StagedSecurityDelegate(
+        dialogHandler = this,
+        stagedWalletSecurityManager = stagedWalletSecurityManager,
+        resourceManager = resourceManager,
+        tariNavigator = tariNavigator,
+    )
+
     private val _uiState = MutableStateFlow(
         HomeOverviewModel.UiState(
             avatarEmoji = corePrefRepository.walletAddress.coreKeyEmojis.extractEmojis().take(1).joinToString(""),
@@ -106,15 +103,11 @@ class HomeOverviewViewModel : CommonViewModel() {
             }
         }
 
-        collectFlow(balanceStateHandler.balanceState) { balanceInfo ->
-            _uiState.update { it.copy(balance = balanceInfo) }
-
-            handleStagedSecurity(balanceInfo)
-        }
+        collectFlow(balanceStateHandler.balanceState) { updateBalance(it) }
 
         doOnWalletRunning {
             doOnWalletServiceConnected {
-                _uiState.update { it.copy(balance = walletManager.requireWalletInstance.getBalance()) }
+                updateBalance(walletManager.requireWalletInstance.getBalance())
                 runCatching { refreshAllData() }
             }
         }
@@ -154,6 +147,11 @@ class HomeOverviewViewModel : CommonViewModel() {
 
     fun handleDeeplink(context: Activity, deepLink: DeepLink) {
         deeplinkManager.execute(context, deepLink)
+    }
+
+    private fun updateBalance(balanceInfo: BalanceInfo) {
+        _uiState.update { it.copy(balance = balanceInfo) }
+        stagedSecurityDelegate.handleStagedSecurity(balanceInfo)
     }
 
     private fun updateList() {
@@ -215,123 +213,6 @@ class HomeOverviewViewModel : CommonViewModel() {
         showSimpleDialog(
             title = resourceManager.getString(error_node_unreachable_title),
             description = resourceManager.getString(error_node_unreachable_description),
-        )
-    }
-
-    /**
-     * Show staged security popups
-     */
-
-    private fun handleStagedSecurity(balanceInfo: BalanceInfo) {
-        stagedWalletSecurityManager.handleBalanceChange(balanceInfo)
-            .takeIfIs<StagedSecurityEffect.ShowStagedSecurityPopUp>()
-            ?.let { effect ->
-                when (effect.stage) {
-                    WalletSecurityStage.Stage1A -> showStagePopUp1A()
-                    WalletSecurityStage.Stage1B -> showStagePopUp1B()
-                    WalletSecurityStage.Stage2 -> showStagePopUp2()
-                    WalletSecurityStage.Stage3 -> showStagePopUp3()
-                }
-            }
-    }
-
-    private fun showStagePopUp1A() {
-        showPopup(
-            stage = BackupOnboardingArgs.StageOne(resourceManager, this::openStage1),
-            titleEmoji = resourceManager.getString(R.string.staged_wallet_security_stages_1a_title),
-            title = resourceManager.getString(R.string.staged_wallet_security_stages_1a_subtitle),
-            body = null,
-            positiveButtonTitle = resourceManager.getString(R.string.staged_wallet_security_stages_1a_buttons_positive),
-            bodyHtml = HtmlHelper.getSpannedText(resourceManager.getString(R.string.staged_wallet_security_stages_1a_message)),
-            positiveAction = { openStage1() },
-        )
-    }
-
-    private fun showStagePopUp1B() {
-        showPopup(
-            stage = BackupOnboardingArgs.StageTwo(resourceManager, this::openStage1B),
-            titleEmoji = resourceManager.getString(R.string.staged_wallet_security_stages_1b_title),
-            title = resourceManager.getString(R.string.staged_wallet_security_stages_1b_subtitle),
-            body = resourceManager.getString(R.string.staged_wallet_security_stages_1b_message),
-            positiveButtonTitle = resourceManager.getString(R.string.staged_wallet_security_stages_1b_buttons_positive),
-            positiveAction = { openStage1() },
-        )
-    }
-
-    private fun showStagePopUp2() {
-        showPopup(
-            stage = BackupOnboardingArgs.StageThree(resourceManager, this::openStage2),
-            titleEmoji = resourceManager.getString(R.string.staged_wallet_security_stages_2_title),
-            title = resourceManager.getString(R.string.staged_wallet_security_stages_2_subtitle),
-            body = resourceManager.getString(R.string.staged_wallet_security_stages_2_message),
-            positiveButtonTitle = resourceManager.getString(R.string.staged_wallet_security_stages_2_buttons_positive),
-            positiveAction = { openStage2() },
-        )
-    }
-
-    private fun showStagePopUp3() {
-        showPopup(
-            stage = BackupOnboardingArgs.StageFour(resourceManager, this::openStage3),
-            titleEmoji = resourceManager.getString(R.string.staged_wallet_security_stages_3_title),
-            title = resourceManager.getString(R.string.staged_wallet_security_stages_3_subtitle),
-            body = resourceManager.getString(R.string.staged_wallet_security_stages_3_message),
-            positiveButtonTitle = resourceManager.getString(R.string.staged_wallet_security_stages_3_buttons_positive),
-            positiveAction = { openStage3() },
-        )
-    }
-
-    private fun openStage1() {
-        hideDialog()
-        tariNavigator.let {
-            it.toAllSettings()
-            it.toBackupSettings(false)
-            it.toWalletBackupWithRecoveryPhrase()
-        }
-    }
-
-    private fun openStage1B() {
-        hideDialog()
-        tariNavigator.let {
-            it.toAllSettings()
-            it.toBackupSettings(true)
-        }
-    }
-
-    private fun openStage2() {
-        hideDialog()
-        tariNavigator.let {
-            it.toAllSettings()
-            it.toBackupSettings(false)
-            it.toChangePassword()
-        }
-    }
-
-    private fun openStage3() {
-        hideDialog()
-        //todo for future
-    }
-
-    private fun showPopup(
-        stage: BackupOnboardingArgs,
-        titleEmoji: String,
-        title: String,
-        body: String?,
-        positiveButtonTitle: String,
-        bodyHtml: Spanned? = null,
-        positiveAction: () -> Unit = {},
-    ) {
-        showModularDialog(
-            SecurityStageHeadModule(titleEmoji, title) { showBackupInfo(stage) },
-            BodyModule(body, bodyHtml?.let { SpannableString(it) }),
-            ButtonModule(positiveButtonTitle, ButtonStyle.Normal) { positiveAction.invoke() },
-            ButtonModule(resourceManager.getString(R.string.staged_wallet_security_buttons_remind_me_later), ButtonStyle.Close),
-        )
-    }
-
-    private fun showBackupInfo(stage: BackupOnboardingArgs) {
-        showModularDialog(
-            BackupOnboardingFlowItemModule(stage),
-            SpaceModule(20),
         )
     }
